Исправлены заменчания наставника
