+++ src/presenter/info.js
@@ -3,7 +3,6 @@ import {render, RenderPosition, remove, replace} from '../utils/render';
 import {sortDefault} from '../utils/sort';
 import {filter} from '../utils/filter';
 
-
 export default class InfoPresenter {
   constructor(container, pointsModel, filterModel) {
     this._container = container;
