+++ src/view/menu.js
@@ -47,5 +47,4 @@ export default class Menu extends AbstractView {
       item.classList.add(`trip-tabs__btn--active`);
     }
   }
-
 }
